rq2-rdf/pyproject.toml
[project]
name = "rq2-rdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.16",
    "faiss-cpu>=1.10.0",
    "gensim>=4.3.3",
    "nest-asyncio>=1.6.0",
    "pydantic>=2.11.3",
    "pyrdf2vec>=0.2.3",
    "rdflib>=7.1.4",
    "requests>=2.32.3",
    "unidecode>=1.3.8",
]
